Web-App to prevent road accidents
Open Datasets provided on the Govt. portal. This was a hackathon organized under the Digital India initiative by the Govt. of India. We analysed the data and found out that 80% of accidents are caused due to drivers, s o w e built a prototype that will do real-time image processing to check whether the driver is loosing is attention while driving and will alert him.